refactor: remove compilation warnings
