+++ src/app/enum/preserved/framework.enum.ts
