+++ src/app/interfaces/preserved/base.ts
