+++ src/app/interfaces/preserved/size.schema.ts
