--- src/app/services/forms/list-form.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ListFormService {
-
-  constructor() { }
-}
