+++ tsconfig.app.json
@@ -12,9 +12,6 @@
     "src/main.ts",
     "src/polyfills.ts"
   ],
-  "include": [
-    "src/**/*.ts"
-  ],
   "exclude": [
     "src/test.ts",
     "src/**/*.spec.ts"
